refactor: 전체적인 프로젝트 리팩토링 미사용 메서드 제거 미사용 테스트 메서드 제거 config 패키지에 config 와 관련된 클래스 이동
