+++ src/main/java/racingcar/domain/Car.java
@@ -3,7 +3,10 @@ package racingcar.domain;
 public interface Car {
 
     public int getPosition();
+
     public String getName();
+
     public void move(int seed);
+
     public String toString();
 }
