+++ src/test/java/racingcar/domain/RacingCarGameTest.java
@@ -8,8 +8,8 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import racingcar.AppConfig;
-import racingcar.AppConfigImpl;
+import racingcar.config.AppConfig;
+import racingcar.config.AppConfigImpl;
 import racingcar.generator.RandomNumberGenerator;
 
 public class RacingCarGameTest {
@@ -62,8 +62,8 @@ public class RacingCarGameTest {
     }
 
     @Test
-    void getCarsSizeTest(){
-        String[] carNames = new String[]{"pobi", "crong","honux"};
+    void getCarsSizeTest() {
+        String[] carNames = new String[]{"pobi", "crong", "honux"};
         racingCarGame = new RacingCarGameImpl(appConfig, carNames);
 
         assertThat(racingCarGame.getCarsSize()).isEqualTo(3);
@@ -71,8 +71,8 @@ public class RacingCarGameTest {
 
 
     @Test
-    public void raceTest(){
-        String[] carNames = new String[]{"pobi", "crong","honux"};
+    public void raceTest() {
+        String[] carNames = new String[]{"pobi", "crong", "honux"};
         racingCarGame = new RacingCarGameImpl(appConfig, carNames);
 
         racingCarGame.race();
@@ -88,25 +88,4 @@ public class RacingCarGameTest {
                 .getPosition()).isEqualTo(0);
 
     }
-//    @Test
-//    public void runTest() throws NoSuchFieldException, IllegalAccessException {
-//        Car car1 = CarFactory.createCar("pobi");
-//        Car car2 = CarFactory.createCar("crong");
-//        Car car3 = CarFactory.createCar("honux");
-//
-//        racingCarGame.add(car1, car2, car3);
-//        racingCarGame.run(5);
-//
-//        Field cars = RacingCarGameImpl.class.getDeclaredField("cars");
-//        cars.setAccessible(true);
-//        ArrayList<Car> racingCars = (ArrayList<Car>) cars.get(racingCarGame);
-//
-//        assertThat(racingCars.get(0)
-//                .getPosition()).isEqualTo(2);
-//        assertThat(racingCars.get(1)
-//                .getPosition()).isEqualTo(3);
-//        assertThat(racingCars.get(2)
-//                .getPosition()).isEqualTo(2);
-//    }
-
 }
